- changed price attribute in MenuItem from integer to decimal (10,2) - added checking types in all functions in menuLogic - changed validate_price, validate_name and added parse_money - added parse_money to functions in menuRouter

+++ source/backend/models/MenuItem.js
@@ -8,7 +8,7 @@ export const MenuItem = sequelize.define("MenuItem", {
     itemID:      { type: DataTypes.INTEGER, autoIncrement: true, primaryKey: true},
     restID:      { type: DataTypes.INTEGER, allowNull: false }, //Restaurant / Foreign Key
     name:        { type: DataTypes.STRING,  allowNull: false },
-    price:       { type: DataTypes.INTEGER, allowNull: false },
+    price:       { type: DataTypes.DECIMAL(10,2), allowNull: false },
     description: { type: DataTypes.TEXT },
     category:    { type: DataTypes.STRING },
 });
